:sparkles: add config.h
